// src/app/models/faturamento/faturamento-mensal-produto.ts
export class FaturamentoMensalProduto{
    cdCliente: number | undefined;
    nomeCliente: string | undefined;
    cdProduto: number | undefined;
    nomeProduto: any;
    cdBeneficiamento: number | undefined;
    nomeBeneficiamento: string | undefined;
    valor: number  = 0;
    area: number  = 0;
    quantidade: number = 0;
}